Change HW version to 4.12. Working "clone" of app_adc as custom app.
